refactor, move loading skeleton to separate file
